If container class is parsing late lexed methods, treat deferred parsed nested classes as non-nested
In such a case the nested class can access the declarations of the container.
If we don't do that the nested classes late parsing of things like member initializers will be deferred and
a lexed method from the container that triggered that late parsing of the nested tag will error out due to
unparsed member initializer.